feat(errorHandling): Stop logging
Logging and throwing and exception is considered
an anti-pattern by many.

* https://community.oracle.com/docs/DOC-983543#logAndThrow
* http://stackoverflow.com/questions/6639963/
why-is-log-and-throw-condsidered-an-anti-pattern

But at the same time, catching an exception too far away from where it
was launched can be problematic if you want to add additional
information to help debugging. Or in other case remove sensitive
infromation before logging.

To solve this I propose that each function is responsible for adding
meta information or removing sensitive data and passes it higher up
the stack without logging.

This avoids unnecessary duplicate logs, ensure that an exception is
logged only if the catcher deems it necessary and it can contain all
the information necessary for debugging and nothing sensitive.